Changed which ttyUSB device we are using from ttyUSB0 --> ttyUSB.RS485
Because this is a fixed name that will always point to which ever device
the name the RS485 USB device is using. This is so that if we connect a
different serial device we won't have conflicts. See this [stackoverflow][1]
answer for how to set it up.

The confi I used to set this up was

/bin/cat <<EOM > /etc/udev/rules.d/10-local.rules
SUBSYSTEM=="tty", ATTRS{product}=="USB-RS485 Cable", ATTRS{idProduct}=="6001", ATTRS{manufacturer}=="FTDI", SYMLINK+="ttyUSB.RS485"
EOM

[1]: https://unix.stackexchange.com/a/66916/33517